requirements.txt
paho-mqtt==1.6.1
prettytable==3.3.0
